Repositories/python/python-microservices-grpc/requirements.txt
flask==2.3.3
grpcio==1.60.0
grpcio-tools==1.60.0
